Add project structure with JS-modules
