+++ js/map/map-init.js
@@ -1,0 +1,3 @@
+const isMapLoaded = () => false;
+
+export { isMapLoaded };
