+++ package.json
@@ -18,7 +18,7 @@
   "license": "MIT",
   "dependencies": {},
   "scripts": {
-    "start": "browser-sync start --server --no-ui --files \"js/**/*.js\"",
+    "start": "browser-sync start --server --no-ui --files \"js/**/*.js, *.css, *.html\"",
     "test": "eslint js/"
   },
   "engines": {
